refactor(app): refactor file structure and Prisma
