--- src/DAL/UserRepository/UserRepository.ts
@@ -1,3 +1,0 @@
-import { PrismaService } from '../prisma/prisma.service';
-
-export default class UserRepository extends PrismaService {}
