--- src/DAL/UserRepository/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserRepository';
