--- src/factories/AbstractDto.ts
@@ -1,5 +1,0 @@
-export abstract class AbstractDto {
-  constructor(input: unknown) {
-    Object.assign(this, input);
-  }
-}
